Got m8r running.
This will now be a simulation on Mac. Eventually it will have 3 TCP ports. One for a shell, one for log output, and one for simulated input and output. The first two will be the same on the device version. The third is only on the m8r command line simulation.

+++ mac/MacTaskManager.cpp
@@ -45,7 +45,7 @@ MacTaskManager::MacTaskManager()
 {
     _eventThread = new std::thread([this]{
         while (true) {
-            int32_t now = SystemInterface::currentMicroseconds() / 1000;
+            int32_t now = static_cast<int32_t>(SystemInterface::currentMicroseconds() / 1000);
             {
                 std::unique_lock<std::mutex> lock(_eventMutex);
                 if (empty()) {
